# matchtree: call prepare on symbolRegexpMatchTree subtree 