--- a/cmd/zoekt/main.go
+++ b/cmd/zoekt/main.go
@@ -130,6 +130,24 @@ func startFullProfile(path string, duration time.Duration) func() bool {
 	})
 }
 
+// experimental support for symbol queries. We just convert substring queries
+// into symbol queries. Needs to run after query.ExpandFileContent
+func toSymbolQuery(q query.Q) query.Q {
+	return query.Map(q, func(q query.Q) query.Q {
+		switch s := q.(type) {
+		case *query.Substring:
+			if s.Content {
+				return &query.Symbol{Expr: s}
+			}
+		case *query.Regexp:
+			if s.Content {
+				return &query.Symbol{Expr: s}
+			}
+		}
+		return q
+	})
+}
+
 func main() {
 	shard := flag.String("shard", "", "search in a specific shard")
 	index := flag.String("index_dir",
@@ -141,6 +159,7 @@ func main() {
 	verbose := flag.Bool("v", false, "print some background data")
 	withRepo := flag.Bool("r", false, "print the repo before the file name")
 	list := flag.Bool("l", false, "print matching filenames only")
+	sym := flag.Bool("sym", false, "do experimental symbol search")
 
 	flag.Usage = func() {
 		name := os.Args[0]
@@ -170,18 +189,23 @@ func main() {
 		log.Fatal(err)
 	}
 
-	query, err := query.Parse(pat)
+	q, err := query.Parse(pat)
 	if err != nil {
 		log.Fatal(err)
 	}
+	q = query.Map(q, query.ExpandFileContent)
+	if *sym {
+		q = toSymbolQuery(q)
+	}
+	q = query.Simplify(q)
 	if *verbose {
-		log.Println("query:", query)
+		log.Println("query:", q)
 	}
 
 	sOpts := zoekt.SearchOptions{
 		DebugScore: *debug,
 	}
-	sres, err := searcher.Search(context.Background(), query, &sOpts)
+	sres, err := searcher.Search(context.Background(), q, &sOpts)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -189,10 +213,10 @@ func main() {
 	// If profiling, do it another time so we measure with
 	// warm caches.
 	for run := startCPUProfile(*cpuProfile, *profileTime); run(); {
-		sres, _ = searcher.Search(context.Background(), query, &sOpts)
+		sres, _ = searcher.Search(context.Background(), q, &sOpts)
 	}
 	for run := startFullProfile(*fullProfile, *profileTime); run(); {
-		sres, _ = searcher.Search(context.Background(), query, &sOpts)
+		sres, _ = searcher.Search(context.Background(), q, &sOpts)
 	}
 
 	displayMatches(sres.Files, pat, *withRepo, *list)