@keegancsmith 2023-11-09
Indeed, this matchtree code needs some updates to be better understood. It took me a while to wrap my head around it back in the day, and debugging with you yesterday made me realise I didn't fully understand it.

> Turns out, without `prepare()`, `nextDoc()` never advances. It seems that `nextDoc()` works functionally like a `peek()` method.

Indeed, it is like a peek method. However, ngramDocIterator does mutate on calls to nextDoc. This is making me slightly uncomfortable here to be honest, it feels like there has to be a bug lurking here.

I think it should never mutate state. I think that also explains why andMatchTree is implemented that way.

> The only reason this ever worked at all was because of [these couple of lines](https://sourcegraph.com/github.com/sourcegraph/zoekt@973c5bcade007e2a1127837e720c3f02af461dbf/-/blob/eval.go?L237-239), which ensure that we _always_ make forward progress. But, that forward progress is one doc at a time, which means we're executing our regex on every single document rather than making use of the index. Thus the dramatic slowness.

aah, great find. I was wondering how this worked at all before. Thanks for the deep dive.