@keegancsmith 2023-11-09
> Great catch @camdencheek ! @keegancsmith is there a way to add test coverage for this fix so we make sure to never regress? It doesn't look easy to unit test, but maybe we could have an integration test like `FilesConsidered` is less than the total number of files?

Yes, I think a test which asserts on stats would be the way to go. I'll see how easy that is to do now, I think it may be straightforward.