@keegancsmith 2023-11-09
@jtibshirani see https://github.com/sourcegraph/zoekt/pull/685/commits/6505dee679f452529c7d2484058c6da6aca9339e for the test case.